Move InOut and ShellAction to controller layer based on TA feedback
