+++ src/main/java/controllers/InOut.java
@@ -1,4 +1,4 @@
-package entities;
+package controllers;
 
 import java.io.IOException;
