+++ src/main/java/controllers/RecipeAppController.java
@@ -1,6 +1,5 @@
 package controllers;
 
-import entities.*;
 import use_cases.LoginManager;
 
 import java.io.IOException;
